status: compute mcoImagePullSpec only once
